+++ tests/games/stendhal/server/maps/quests/SuntanCreamForZaraTest.java
@@ -141,7 +141,7 @@ public class SuntanCreamForZaraTest {
 		// [10:02] Changed the state of quest 'pamela_mix_cream' from '1;suntan cream;1288519190459' to '1;suntan cream;0'
 		player.setQuest("pamela_mix_cream", "1;suntan cream;0");
 
-		final int xp = player.getXP();
+		final long xp = player.getXP();
 		en.step(player, "hi");
 		assertEquals("Welcome back! I'm done with your order. Here you have the suntan cream.", getReply(npc));
 		// [10:02] kymara earns 1 experience point.
@@ -172,7 +172,7 @@ public class SuntanCreamForZaraTest {
 		en.step(player, "bye");
 		assertEquals("I hope to see you soon!", getReply(npc));
 
-		final int xp = player.getXP();
+		final long xp = player.getXP();
 		final double karma = player.getKarma();
 
 		en.step(player, "hi");
